requirements.txt
keyboard==0.13.5
openai_whisper==20240930
PyAudio==0.2.14
torch==2.5.1+cu124
